// COMPLEMENTARIOS-2P/ComplementarioCSR/Complementario3/src/interfaces/iregistroviajes.ts
// Generated by https://quicktype.io

export interface IResRegistroviajes {
    total: number;
    datos: Registroviajes[];
}

export interface Registroviajes{
    _id?:       string;
    IdBus:     IDBus;
    IdUsuario: IDUsuario;
    Estado?:    boolean;
    _v?:       number;
}
export interface IDBus {
    _id?:           string;
    Placa:         string;
    NumeroBus:     string;
    Capacidad:     number;
    CedChofer:     string;
    NombresChofer: string;
    Estado?:        boolean;
    _v?:           number;
    Contraseña:    string;
}

export interface IDUsuario {
    _id?:        string;
    Cedula:     string;
    Nombres:    string;
    CntViajes:  number;
    Estado?:     boolean;
    _v?:        number;
    Contraseña: string;
    TypeUser:   string;
}